pyproject.toml
[tool.poetry]
authors = ["Evžen Wybitul <wybitul.evzen@gmail.com>"]
description = ""
name = "bioinformatics-toolbox"
version = "0.1.0"

[tool.poetry.dependencies]
Sphinx = "^4.0.1"
biopython = "^1.78"
ipython = "^7.23.1"
jedi = "0.17.2"
jupyter = "^1.0.0"
numpy = "^1.20.3"
python = "3.9.4"
scipy = "^1.6.3"

[tool.poetry.dev-dependencies]
black = {version = "^21.5b1", allow-prereleases = true}

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
